more refactoring and corrections
